+++ Data/SBV/Examples/Misc/Floating.hs
@@ -14,6 +14,8 @@
 -- the presence of @NaN@ is always something to look out for.
 -----------------------------------------------------------------------------
 
+{-# LANGUAGE ScopedTypeVariables #-}
+
 module Data.SBV.Examples.Misc.Floating where
 
 import Data.SBV
@@ -109,7 +111,7 @@ nonZeroAddition = prove $ do [a, b] <- sFloats ["a", "b"]
 -- * FP multiplicative inverses may not exist
 -----------------------------------------------------------------------------
 
--- | The last example illustrates that @a * (1/a)@ does not necessarily equal @1@. Again,
+-- | This example illustrates that @a * (1/a)@ does not necessarily equal @1@. Again,
 -- we protect against division by @0@ and @NaN@/@Infinity@.
 --
 -- We have:
@@ -120,7 +122,7 @@ nonZeroAddition = prove $ do [a, b] <- sFloats ["a", "b"]
 --
 -- Indeed, we have:
 --
--- >>> let a =  1.2354518252390238e308 :: Double
+-- >>> let a = 1.2354518252390238e308 :: Double
 -- >>> a * (1/a)
 -- 0.9999999999999998
 multInverse :: IO ThmResult
@@ -128,3 +130,68 @@ multInverse = prove $ do a <- sDouble "a"
                          constrain $ isFPPoint a
                          constrain $ isFPPoint (1/a)
                          return $ a * (1/a) .== 1
+
+-----------------------------------------------------------------------------
+-- * Effect of rounding modes
+-----------------------------------------------------------------------------
+
+-- | One interesting aspect of floating-point is that the chosen rounding-mode
+-- can effect the results of a computation if the exact result cannot be precisely
+-- represented. SBV exports the functions 'fpAdd', 'fpSub', 'fpMul', 'fpDiv', 'fpFMA'
+-- and 'fpSqrt' which allows users to specify the IEEE supported 'RoundingMode' for
+-- the operation. (Also see the class 'RoundingFloat'.) This example illustrates how SBV
+-- can be used to find rounding-modes where, for instance, addition can produce different
+-- results. We have:
+--
+-- >>> roundingAdd
+-- Satisfiable. Model:
+--   rm = RoundTowardPositive :: RoundingMode
+--   x = 1.7014118e38 :: SFloat
+--   y = 1.1754942e-38 :: SFloat
+--
+-- Unfortunately we can't directly validate this result at the Haskell level, as Haskell only supports
+-- 'RoundNearestTiesToEven'. We have:
+--
+-- >>> (1.7014118e38 + 1.1754942e-38) :: Float
+-- 1.7014118e38
+--
+-- Note that result is identical to the first argument. But with a 'RoundTowardPositive', we would
+-- get the result @1.701412e38@. While we cannot directly see this from within Haskell, we can
+-- use SBV to provide us with that result thusly:
+--
+-- >>> sat $ \x -> x .== fpAdd (literal RoundTowardPositive) (1.7014118e38::SFloat)  (1.1754942e-38::SFloat)
+-- Satisfiable. Model:
+--   s0 = 1.701412e38 :: SFloat
+--
+-- We can see why these two resuls are different if we treat these values as arbitrary
+-- precision reals, as represented by the 'SReal' type:
+--
+-- >>> let x = 1.7014118e38 :: SReal
+-- >>> let y = 1.1754942e-38 :: SReal
+-- >>> x
+-- 170141180000000000000000000000000000000.0 :: SReal
+-- >>> y
+-- 0.000000000000000000000000000000000000011754942 :: SReal
+-- >>> x + y
+-- 170141180000000000000000000000000000000.000000000000000000000000000000000000011754942 :: SReal
+--
+-- When we do 'RoundNearestTiesToEven', the entire suffix falls off, as it happens that the infinitely
+-- precise result is closer to the value of @x@. But when we use 'RoundTowardPositive', we reach
+-- for the next representable number, which happens to be @1.701412e38@. You might wonder why not
+-- @1.7014119e38@? Because that number is not precisely representable as a 'Float':
+--
+-- >>> 1.7014119e38:: Float
+-- 1.7014118e38
+--
+-- But @1.701412e38@ is:
+--
+-- >>> 1.701412e38 :: Float
+-- 1.701412e38
+--
+-- Floating point representation and semantics is indeed a thorny subject, <https://ece.uwaterloo.ca/~dwharder/NumericalAnalysis/02Numerics/Double/paper.pdf> happens to be an excellent guide, however.
+roundingAdd :: IO SatResult
+roundingAdd = sat $ do m :: SRoundingMode <- free "rm"
+                       x <- sFloat "x"
+                       y <- sFloat "y"
+                       constrain $ m ./= literal RoundNearestTiesToEven
+                       return $ fpAdd m x y ./= x + y