+++ Data/SBV/Provers/Prover.hs
@@ -26,7 +26,7 @@ module Data.SBV.Provers.Prover (
        , isVacuous, isVacuousWith
        , SatModel(..), Modelable(..), displayModels, extractModels
        , getModelDictionaries, getModelValues, getModelUninterpretedValues
-       , boolector, cvc4, yices, z3, mathSAT, defaultSMTCfg
+       , boolector, cvc4, yices, z3, mathSAT, abc, defaultSMTCfg
        , compileToSMTLib, generateSMTBenchmarks
        , isSBranchFeasibleInState
        , isConditionSatisfiable
@@ -52,6 +52,7 @@ import qualified Data.SBV.Provers.CVC4       as CVC4
 import qualified Data.SBV.Provers.Yices      as Yices
 import qualified Data.SBV.Provers.Z3         as Z3
 import qualified Data.SBV.Provers.MathSAT    as MathSAT
+import qualified Data.SBV.Provers.ABC        as ABC
 import Data.SBV.Utils.TDiff
 
 mkConfig :: SMTSolver -> Bool -> [String] -> SMTConfig
@@ -89,6 +90,10 @@ z3 = mkConfig Z3.z3 True ["(set-option :smt.mbqi true) ; use model based quantif
 -- | Default configuration for the MathSAT SMT solver
 mathSAT :: SMTConfig
 mathSAT = mkConfig MathSAT.mathSAT True []
+
+-- | Default configuration for the ABC synthesis and verification tool.
+abc :: SMTConfig
+abc = mkConfig ABC.abc True []
 
 -- | The default solver used by SBV. This is currently set to z3.
 defaultSMTCfg :: SMTConfig
@@ -371,7 +376,9 @@ allSatWith config p = do
         let converter = if useSMTLib2 config then toSMTLib2 else toSMTLib1
         msg "Checking Satisfiability, all solutions.."
         sbvPgm@(qinps, _, _, ki, _) <- simulate converter config True [] p
-        let usorts = [s | KUserSort s _ <- Set.toList ki]
+        let usorts = [s | us@(KUserSort s _) <- Set.toList ki, isFree us]
+                where isFree (KUserSort _ (Left _, _)) = True
+                      isFree _                         = False
         unless (null usorts) $ msg $  "SBV.allSat: Uninterpreted sorts present: " ++ unwords usorts
                                    ++ "\n               SBV will use equivalence classes to generate all-satisfying instances."
         results <- unsafeInterleaveIO $ go sbvPgm (1::Int) []