+++ Data/SBV/Provers/SExpr.hs
@@ -11,16 +11,23 @@
 
 module Data.SBV.Provers.SExpr where
 
+import Data.Bits            (setBit, testBit)
+import Data.Word            (Word32, Word64)
 import Data.Char            (isDigit, ord)
 import Data.List            (isPrefixOf)
+import Data.Maybe           (fromMaybe, listToMaybe)
 import Numeric              (readInt, readDec, readHex, fromRat)
 
 import Data.SBV.BitVectors.AlgReals
-import Data.SBV.BitVectors.Data (nan, infinity)
+import Data.SBV.BitVectors.Data (nan, infinity, RoundingMode(..))
+
+-- Needed for conversion from SMTLib triple-IEEE formats to float/double
+import qualified Foreign          as F
+import qualified System.IO.Unsafe as U (unsafePerformIO)  -- Only used safely!
 
 -- | ADT S-Expression format, suitable for representing get-model output of SMT-Lib
 data SExpr = ECon    String
-           | ENum    Integer
+           | ENum    (Integer, Maybe Int)  -- Second argument is how wide the field was in bits, if known. Useful in FP parsing.
            | EReal   AlgReal
            | EFloat  Float
            | EDouble Double
@@ -55,44 +62,57 @@ parseSExpr inp = do (sexp, extras) <- parse inpToks
                                        parseApp r (f : sofar)
         parseApp (tok:toks) sofar = do t <- pTok tok
                                        parseApp toks (t : sofar)
-        pTok "false"              = return $ ENum 0
-        pTok "true"               = return $ ENum 1
-        pTok ('0':'b':r)          = mkNum $ readInt 2 (`elem` "01") (\c -> ord c - ord '0') r
-        pTok ('b':'v':r)          = mkNum $ readDec (takeWhile (/= '[') r)
-        pTok ('#':'b':r)          = mkNum $ readInt 2 (`elem` "01") (\c -> ord c - ord '0') r
-        pTok ('#':'x':r)          = mkNum $ readHex r
+        pTok "false"              = return $ ENum (0, Nothing)
+        pTok "true"               = return $ ENum (1, Nothing)
+        pTok ('0':'b':r)          = mkNum (Just (length r))     $ readInt 2 (`elem` "01") (\c -> ord c - ord '0') r
+        pTok ('b':'v':r)          = mkNum Nothing               $ readDec (takeWhile (/= '[') r)
+        pTok ('#':'b':r)          = mkNum (Just (length r))     $ readInt 2 (`elem` "01") (\c -> ord c - ord '0') r
+        pTok ('#':'x':r)          = mkNum (Just (4 * length r)) $ readHex r
         pTok n
           | not (null n) && isDigit (head n)
           = if '.' `elem` n then getReal n
-            else mkNum  $ readDec n
-        pTok n                 = return $ ECon n
-        mkNum [(n, "")] = return $ ENum n
-        mkNum _         = die "cannot read number"
+            else mkNum Nothing $ readDec n
+        pTok n                 = return $ ECon (constantMap n)
+        mkNum l [(n, "")] = return $ ENum (n, l)
+        mkNum _ _         = die "cannot read number"
         getReal n = return $ EReal $ mkPolyReal (Left (exact, n'))
           where exact = not ("?" `isPrefixOf` reverse n)
                 n' | exact = n
                    | True  = init n
         -- simplify numbers and root-obj values
         cvt (EApp [ECon "/", EReal a, EReal b])                    = return $ EReal (a / b)
-        cvt (EApp [ECon "/", EReal a, ENum  b])                    = return $ EReal (a             / fromInteger b)
-        cvt (EApp [ECon "/", ENum  a, EReal b])                    = return $ EReal (fromInteger a /             b)
-        cvt (EApp [ECon "/", ENum  a, ENum  b])                    = return $ EReal (fromInteger a / fromInteger b)
+        cvt (EApp [ECon "/", EReal a, ENum  b])                    = return $ EReal (a                   / fromInteger (fst b))
+        cvt (EApp [ECon "/", ENum  a, EReal b])                    = return $ EReal (fromInteger (fst a) /             b      )
+        cvt (EApp [ECon "/", ENum  a, ENum  b])                    = return $ EReal (fromInteger (fst a) / fromInteger (fst b))
         cvt (EApp [ECon "-", EReal a])                             = return $ EReal (-a)
-        cvt (EApp [ECon "-", ENum a])                              = return $ ENum  (-a)
+        cvt (EApp [ECon "-", ENum a])                              = return $ ENum  (-(fst a), snd a)
         -- bit-vector value as CVC4 prints: (_ bv0 16) for instance
         cvt (EApp [ECon "_", ENum a, ENum _b])                     = return $ ENum a
         cvt (EApp [ECon "root-obj", EApp (ECon "+":trms), ENum k]) = do ts <- mapM getCoeff trms
-                                                                        return $ EReal $ mkPolyReal (Right (k, ts))
-        cvt (EApp [ECon "as", n, EApp [ECon "_", ECon "FloatingPoint", ENum 11, ENum 53]]) = getDouble n
-        cvt (EApp [ECon "as", n, EApp [ECon "_", ECon "FloatingPoint", ENum  8, ENum 24]]) = getFloat  n
-        cvt (EApp [ECon "as", n, ECon "Float64"])                                          = getDouble n
-        cvt (EApp [ECon "as", n, ECon "Float32"])                                          = getFloat  n
-        cvt x                                                      = return x
-        getCoeff (EApp [ECon "*", ENum k, EApp [ECon "^", ECon "x", ENum p]]) = return (k, p)  -- kx^p
-        getCoeff (EApp [ECon "*", ENum k,                 ECon "x"        ] ) = return (k, 1)  -- kx
-        getCoeff (                        EApp [ECon "^", ECon "x", ENum p] ) = return (1, p)  --  x^p
-        getCoeff (                                        ECon "x"          ) = return (1, 1)  --  x
-        getCoeff (                ENum k                                    ) = return (k, 0)  -- k
+                                                                        return $ EReal $ mkPolyReal (Right (fst k, ts))
+        cvt (EApp [ECon "as", n, EApp [ECon "_", ECon "FloatingPoint", ENum (11, _), ENum (53, _)]]) = getDouble n
+        cvt (EApp [ECon "as", n, EApp [ECon "_", ECon "FloatingPoint", ENum ( 8, _), ENum (24, _)]]) = getFloat  n
+        cvt (EApp [ECon "as", n, ECon "Float64"])                                                    = getDouble n
+        cvt (EApp [ECon "as", n, ECon "Float32"])                                                    = getFloat  n
+        -- NB. Note the lengths on the mantissa for the following two are 23/52; not 24/53!
+        cvt (EApp [ECon "fp",    ENum (s, Just 1), ENum ( e, Just 8),  ENum (m, Just 23)])           = return $ EFloat  $ getTripleFloat  s e m
+        cvt (EApp [ECon "fp",    ENum (s, Just 1), ENum ( e, Just 11), ENum (m, Just 52)])           = return $ EDouble $ getTripleDouble s e m
+        cvt (EApp [ECon "_",     ECon "NaN",       ENum ( 8, _),       ENum (24,      _)])           = return $ EFloat  nan
+        cvt (EApp [ECon "_",     ECon "NaN",       ENum (11, _),       ENum (53,      _)])           = return $ EDouble nan
+        cvt (EApp [ECon "_",     ECon "+oo",       ENum ( 8, _),       ENum (24,      _)])           = return $ EFloat  infinity
+        cvt (EApp [ECon "_",     ECon "+oo",       ENum (11, _),       ENum (53,      _)])           = return $ EDouble infinity
+        cvt (EApp [ECon "_",     ECon "-oo",       ENum ( 8, _),       ENum (24,      _)])           = return $ EFloat  (-infinity)
+        cvt (EApp [ECon "_",     ECon "-oo",       ENum (11, _),       ENum (53,      _)])           = return $ EDouble (-infinity)
+        cvt (EApp [ECon "_",     ECon "+zero",     ENum ( 8, _),       ENum (24,      _)])           = return $ EFloat  0
+        cvt (EApp [ECon "_",     ECon "+zero",     ENum (11, _),       ENum (53,      _)])           = return $ EDouble 0
+        cvt (EApp [ECon "_",     ECon "-zero",     ENum ( 8, _),       ENum (24,      _)])           = return $ EFloat  (-0)
+        cvt (EApp [ECon "_",     ECon "-zero",     ENum (11, _),       ENum (53,      _)])           = return $ EDouble (-0)
+        cvt x                                                                                        = return x
+        getCoeff (EApp [ECon "*", ENum k, EApp [ECon "^", ECon "x", ENum p]]) = return (fst k, fst p)  -- kx^p
+        getCoeff (EApp [ECon "*", ENum k,                 ECon "x"        ] ) = return (fst k,     1)  -- kx
+        getCoeff (                        EApp [ECon "^", ECon "x", ENum p] ) = return (    1, fst p)  --  x^p
+        getCoeff (                                        ECon "x"          ) = return (    1,     1)  --  x
+        getCoeff (                ENum k                                    ) = return (fst k,     0)  -- k
         getCoeff x = die $ "Cannot parse a root-obj,\nProcessing term: " ++ show x
         getDouble (ECon s)  = case (s, rdFP (dropWhile (== '+') s)) of
                                 ("plusInfinity",  _     ) -> return $ EDouble infinity
@@ -131,3 +151,32 @@ rdFP s = case break (`elem` "pe") s of
  where rd v = case reads v of
                 [(n, "")] -> Just n
                 _         -> Nothing
+
+-- | Convert an (s, e, m) triple to a float value
+getTripleFloat :: Integer -> Integer -> Integer -> Float
+getTripleFloat s e m = U.unsafePerformIO $ F.alloca $ \buf -> do {F.poke (F.castPtr buf) w32; F.peek buf}
+  where sign      = [s == 1]
+        expt      = [e `testBit` i | i <- [ 7,  6 .. 0]]
+        mantissa  = [m `testBit` i | i <- [22, 21 .. 0]]
+        positions = [i | (i, b) <- zip [31, 30 .. 0] (sign ++ expt ++ mantissa), b]
+        w32       = foldr (flip setBit) (0::Word32) positions
+
+-- | Convert an (s, e, m) triple to a float value
+getTripleDouble :: Integer -> Integer -> Integer -> Double
+getTripleDouble s e m = U.unsafePerformIO $ F.alloca $ \buf -> do {F.poke (F.castPtr buf) w64; F.peek buf}
+  where sign      = [s == 1]
+        expt      = [e `testBit` i | i <- [10,  9 .. 0]]
+        mantissa  = [m `testBit` i | i <- [51, 50 .. 0]]
+        positions = [i | (i, b) <- zip [63, 62 .. 0] (sign ++ expt ++ mantissa), b]
+        w64       = foldr (flip setBit) (0::Word64) positions
+
+-- | Special constants of SMTLib2 and their internal translation. Mainly
+-- rounding modes for now.
+constantMap :: String -> String
+constantMap n = fromMaybe n (listToMaybe [to | (from, to) <- special, n `elem` from])
+ where special = [ (["RNE", "roundNearestTiesToEven"], show RoundNearestTiesToEven)
+                 , (["RNA", "roundNearestTiesToAway"], show RoundNearestTiesToAway)
+                 , (["RTP", "roundTowardPositive"],    show RoundTowardPositive)
+                 , (["RTN", "roundTowardNegative"],    show RoundTowardNegative)
+                 , (["RTZ", "roundTowardZero"],        show RoundTowardZero)
+                 ]